FTVVIZ-976: INT overflow during MP3 streams
[Problem]
Sample counter is 32 bit, and can overflow during a long playback

[Solution]
Change it to long.

[Test]
Checked only short playback, please test long ones:
- using ethernet connection
- stable server
- longer than 13.5 hours depending in Sample/Frame ratio.

See Jira for details

Reviewed-on: https://e-gerrit.labcollab.net/35298
Tested-by: Zsolt Matyas <zmmatyas@amazon.com>

+++ library/src/main/java/com/google/android/exoplayer/extractor/mp3/Mp3Extractor.java
@@ -69,7 +69,7 @@ public final class Mp3Extractor implements Extractor {
 
   private Seeker seeker;
   private long basisTimeUs;
-  private int samplesRead;
+  private long samplesRead; // AMZN_CHANGE_ONELINE: int overflow after 48695 seconds. See FTVVIZ-976
   private int sampleBytesRemaining;
 
   /** Constructs a new {@link Mp3Extractor}. */
